chore(netflix/errors): convert exception handler
* convert the exception handler to typescript.
